#7 Refactor file domain

+++ app/Domains/File/Services/FileService.php
@@ -2,33 +2,43 @@
 
 namespace App\Domains\File\Services;
 
+use App\Domains\File\Enums\FileTypesEnum;
 use App\Domains\File\Events\FileUploadedEvent;
-use App\Domains\File\Repositories\FileRepositoryInterface;
 use App\Domains\File\DTOs\FileDTO;
+use App\Domains\File\Interfaces\FileOwnerInterface;
+use App\Domains\File\Interfaces\FileRepositoryInterface;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Str;
 
 class FileService
 {
+
+  protected string $disk;
+
+  protected string $directory;
+
   public function __construct(
     private FileRepositoryInterface $repo
   ) {
+
   }
 
-  public function upload(UploadedFile $file, mixed $owner): FileDTO
+  public function upload(UploadedFile $file, FileOwnerInterface $owner, FileTypesEnum $type): FileDTO
   {
-    $disk = 'public';
-    $directory = strtolower(class_basename($owner)) . '/' . $owner->id;
-    $filename = now()->timestamp . '.' . $file->extension();
+    $disk = $owner->getStorageDisk();
+    $directory = $owner->getStorageDirectory();
+    $filename = Str::uuid() . '.' . $file->extension();
     $path = Storage::disk($disk)->putFileAs($directory, $file, $filename);
     $dto = new FileDTO(
-      null,
+      id: null,
       disk: $disk,
       path: $path,
       mimeType: $file->getMimeType(),
       size: $file->getSize(),
-      modelId: $owner->id,
-      modelType: get_class($owner)
+      modelId: $owner->getFileKey(),
+      modelType: get_class($owner),
+      type: $type
     );
 
     $saved = $this->repo->save($dto);
@@ -39,6 +49,10 @@ class FileService
 
   public function delete(int $fileId): void
   {
-    $this->repo->delete($fileId);
+    $file = $this->repo->findOne($fileId);
+    if ($file) {
+      Storage::disk($file->disk)->delete($file->path);
+      $this->repo->delete($fileId);
+    }
   }
 }
